test(fakes): split tests/fakes.py into a package; close test sockets (#445 item 5)
tests/fakes.py reached ~6,100 lines. Split into one module per
collaborator — fakes/pipeline_db.py (FakePipelineDB + Source + plan
rows), fakes/beets.py, fakes/slskd.py, fakes/ytmusic.py,
fakes/lookups.py (MB/Discogs + http_error), fakes/cursors.py,
fakes/rows.py (recorder dataclasses), fakes/_shared.py (time helpers).
`from tests.fakes import X` keeps working for every previously
top-level name via __init__ re-exports.

The move is mechanically verified lossless: an AST comparison of the
old file against the package shows every top-level class/function/
assignment present exactly once with an identical body. The single
behavioural edit is the patch target in test_fakes.py
(tests.fakes._utcnow → tests.fakes.pipeline_db._utcnow — the binding
FakePipelineDB actually resolves).

Also kills the ResourceWarning unclosed-socket spam from tests/web:
the harness _get/_post now close their HTTP responses (success and
HTTPError sides), tearDownClass closes the listening socket after
shutdown(), and the four direct urlopen call sites get the same
treatment. Full-suite output now carries zero ResourceWarnings.

Suite 4453 OK, pyright 0 errors.
