+++ tests/fakes/pipeline_db.py
@@ -1,30 +1,21 @@
-"""Lightweight fakes for stateful collaborators.
+"""FakePipelineDB — the in-memory PipelineDB stand-in.
 
-FakePipelineDB records state transitions, log rows, denylist entries, and
-cooldowns in-memory. Use it in orchestration tests to assert domain outcomes
+Records state transitions, log rows, denylist entries, and cooldowns
+in-memory. Use it in orchestration tests to assert domain outcomes
 instead of MagicMock call shapes.
 """
 
 from __future__ import annotations
 
 import copy
-import email.message
 import json
 import time
-import urllib.error
 from contextlib import contextmanager
 from dataclasses import dataclass, field
 from datetime import date, datetime, timedelta, timezone
 from typing import TYPE_CHECKING, Any, Callable, Iterable, Iterator, Optional
-from zoneinfo import ZoneInfo
 import msgspec
 
-# Single source of truth for Perth-local bucketing inside the fake.
-# Mirrors `(first_seen_at AT TIME ZONE 'Australia/Perth')::date` in
-# `lib/pipeline_db/dashboard.py::get_peer_metrics`. Using UTC
-# bucketing instead would silently disagree with prod by 8h at the
-# day boundary.
-_PERTH_TZ = ZoneInfo("Australia/Perth")
 
 if TYPE_CHECKING:
     from lib.quality import CandidateScore
@@ -77,126 +68,15 @@ from lib.search_classification import (
 )
 
 
-def _utcnow() -> datetime:
-    return datetime.now(timezone.utc)
-
-
-_EPOCH = datetime.min.replace(tzinfo=timezone.utc)
-
-
-def _as_datetime(value: Any) -> datetime:
-    """Normalise a timestamp-ish value to an aware ``datetime``.
-
-    Most test rows now carry real datetimes via ``make_request_row``,
-    but older hand-rolled fixtures still use ISO strings. Sorting with a
-    mixed key would raise ``TypeError``; this helper collapses both
-    shapes to a comparable datetime (aware, UTC) and uses ``_EPOCH`` as
-    the sentinel for missing values so ordering stays deterministic.
-    """
-    if value is None:
-        return _EPOCH
-    if isinstance(value, datetime):
-        return value if value.tzinfo else value.replace(tzinfo=timezone.utc)
-    if isinstance(value, str):
-        try:
-            parsed = datetime.fromisoformat(value)
-        except ValueError:
-            return _EPOCH
-        return parsed if parsed.tzinfo else parsed.replace(
-            tzinfo=timezone.utc)
-    return _EPOCH
-
-
-@dataclass
-class DownloadLogRow:
-    """One row in download_log, captured by FakePipelineDB.log_download."""
-    request_id: int
-    outcome: str | None = None
-    soulseek_username: str | None = None
-    filetype: str | None = None
-    beets_distance: float | None = None
-    beets_scenario: str | None = None
-    beets_detail: str | None = None
-    staged_path: str | None = None
-    error_message: str | None = None
-    validation_result: Any = None
-    import_result: Any = None
-    # Auto-assigned monotonic id matching PostgreSQL serial behaviour.
-    id: int = 0
-    # Migration 021: addressing FK to album_quality_evidence(id).
-    candidate_evidence_id: int | None = None
-    # Migration 037 — source discriminator + YT-specific JSONB blob.
-    # ``source`` defaults to ``'slskd'`` matching the production NOT NULL
-    # DEFAULT. ``youtube_metadata`` is NULL unless this row was written
-    # via ``insert_youtube_running``.
-    source: str = "slskd"
-    youtube_metadata: dict[str, Any] | None = None
-    # Auto-populated timestamp matching download_log.created_at.
-    created_at: datetime = field(default_factory=_utcnow)
-    # Catch-all for less commonly asserted fields
-    extra: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class DenylistEntry:
-    """One row in source_denylist."""
-    request_id: int
-    username: str
-    reason: str | None = None
-
-
-@dataclass
-class SearchLogRow:
-    """One row in search_log, captured by FakePipelineDB.log_search."""
-    request_id: int
-    query: str | None = None
-    result_count: int | None = None
-    elapsed_s: float | None = None
-    outcome: str = "error"
-    id: int = 0
-    created_at: datetime = field(default_factory=_utcnow)
-    # Forensic capture (U5 of search-escalation-and-forensics).
-    # ``candidates`` is the JSONB blob persisted by ``log_search`` — the
-    # in-memory representation is the JSON string the production code
-    # would have written via ``msgspec.json.encode``. NULL on error rows.
-    candidates: str | None = None
-    variant: str | None = None
-    final_state: str | None = None
-    browse_time_s: float = 0.0
-    match_time_s: float = 0.0
-    peers_browsed: int = 0
-    peers_browsed_lazy: int = 0
-    fanout_waves: int = 0
-    # U1 persisted-search-plans plan-context fields. All nullable; rows
-    # written via ``log_search`` keep them None to mirror historical /
-    # legacy production rows.
-    plan_id: int | None = None
-    plan_item_id: int | None = None
-    plan_ordinal: int | None = None
-    plan_strategy: str | None = None
-    plan_canonical_query_key: str | None = None
-    plan_repeat_group: str | None = None
-    plan_generator_id: str | None = None
-    execution_stage: str | None = None
-    attempt_consumed: bool | None = None
-    cursor_update_status: str | None = None
-    stale_reason: str | None = None
-    plan_cycle_snapshot: int | None = None
-    # U2 of search-plan-entropy: pre-filter skip count. NOT NULL on
-    # the real column with default 0; mirror that here so test asserts
-    # never see None on the field.
-    pre_filter_skip_count: int = 0
-    # U11 forensics columns (R22-R27). All nullable in production;
-    # default ``None`` here so legacy / pre-attempt rows mirror the
-    # real DB ``NULL`` shape.
-    rejection_reason: str | None = None
-    result_count_uncapped: int | None = None
-    query_token_count: int | None = None
-    query_distinct_token_count: int | None = None
-    expected_track_count: int | None = None
-    matcher_score_top1: float | None = None
-    query_template: str | None = None
-
+from tests.fakes._shared import _EPOCH, _PERTH_TZ, _as_datetime, _utcnow
+from tests.fakes.cursors import FakeCursor
+from tests.fakes.rows import (
+    DenylistEntry,
+    DownloadLogRow,
+    FieldResolutionRow,
+    SearchLogRow,
+    UserCooldownRow,
+)
 
 @dataclass
 class _FakeSearchPlanRow:
@@ -227,744 +107,6 @@ class _FakeSearchPlanItemRow:
     provenance: dict[str, Any] | None = None
 
 
-@dataclass
-class UserCooldownRow:
-    """One row in user_cooldowns, captured by FakePipelineDB.add_cooldown."""
-    username: str
-    cooldown_until: datetime
-    reason: str | None = None
-    created_at: datetime = field(default_factory=_utcnow)
-
-
-@dataclass
-class FieldResolutionRow:
-    """One row in album_request_field_resolutions (migration 030).
-
-    Captured by ``FakePipelineDB.record_field_resolution``. Mirrors the
-    production schema field-for-field so tests can assert against the
-    same row shape PipelineDB returns from ``get_field_resolution``.
-    """
-    request_id: int
-    field_name: str
-    status: str
-    reason_code: str | None
-    attempts: int = 1
-    resolved_at: datetime = field(default_factory=_utcnow)
-    id: int = 0
-
-
-@dataclass
-class EnqueueCall:
-    """One slskd enqueue call captured by FakeSlskdAPI."""
-    username: str
-    files: list[dict[str, Any]]
-
-
-@dataclass
-class CancelDownloadCall:
-    """One slskd cancel_download call captured by FakeSlskdAPI."""
-    username: str
-    id: str
-
-
-class FakeSlskdTransfers:
-    """Stateful fake for the slskd transfers API."""
-
-    def __init__(self, api: "FakeSlskdAPI") -> None:
-        self._api = api
-        self.enqueue_calls: list[EnqueueCall] = []
-        self.get_all_downloads_calls: list[bool] = []
-        self.get_download_calls: list[tuple[str, str]] = []
-        self.get_downloads_calls: list[tuple[tuple[Any, ...], dict[str, Any]]] = []
-        self.cancel_download_calls: list[CancelDownloadCall] = []
-        self.enqueue_result = True
-        self.enqueue_error: Exception | None = None
-        self.get_all_downloads_error: Exception | None = None
-        self.get_download_error: Exception | None = None
-        self.cancel_download_error: Exception | None = None
-        self.cancel_download_result = True
-
-    def enqueue(self, username: str, files: list[dict[str, Any]]) -> bool:
-        self.enqueue_calls.append(EnqueueCall(username, copy.deepcopy(files)))
-        if self.enqueue_error is not None:
-            raise self.enqueue_error
-        return self.enqueue_result
-
-    def get_all_downloads(self, includeRemoved: bool = False) -> list[dict[str, Any]]:
-        self.get_all_downloads_calls.append(includeRemoved)
-        if self.get_all_downloads_error is not None:
-            raise self.get_all_downloads_error
-        return self._api._next_download_snapshot()
-
-    def get_downloads(self, *args: Any, **kwargs: Any) -> list[dict[str, Any]]:
-        self.get_downloads_calls.append((args, copy.deepcopy(kwargs)))
-        return self.get_all_downloads(
-            includeRemoved=bool(kwargs.get("includeRemoved", False)))
-
-    def get_download(self, username: str, id: str) -> dict[str, Any]:
-        self.get_download_calls.append((username, id))
-        if self.get_download_error is not None:
-            raise self.get_download_error
-        transfer = self._api._find_transfer(username, id)
-        if transfer is None:
-            raise KeyError(f"No transfer {id!r} for {username!r}")
-        return transfer
-
-    def cancel_download(self, username: str, id: str) -> bool:
-        self.cancel_download_calls.append(CancelDownloadCall(username, id))
-        if self.cancel_download_error is not None:
-            raise self.cancel_download_error
-        if not self.cancel_download_result:
-            return False
-        self._api.remove_transfer(username=username, id=id)
-        return True
-
-
-class FakeSlskdUsers:
-    """Stateful fake for the slskd users API."""
-
-    def __init__(self) -> None:
-        self.directory_calls: list[tuple[str, str]] = []
-        self.directory_error: Exception | None = None
-        self._directories: dict[tuple[str, str], list[Any]] = {}
-        self._directory_errors: dict[tuple[str, str], Exception] = {}
-        self._directory_delays: dict[tuple[str, str], float] = {}
-        self.status_calls: list[str] = []
-        self._statuses: dict[str, str] = {}
-        self._status_errors: dict[str, Exception] = {}
-        # Optional concurrency probe — set to a callable taking a +/-1 delta to
-        # observe in-flight count (used by fan-out concurrency-cap tests).
-        self.in_flight_probe: Callable[[int], None] | None = None
-
-    def set_directory(
-        self,
-        username: str,
-        directory: str,
-        result: list[Any],
-    ) -> None:
-        self._directories[(username, directory)] = copy.deepcopy(result)
-
-    def set_directory_error(
-        self,
-        username: str,
-        directory: str,
-        error: Exception,
-    ) -> None:
-        self._directory_errors[(username, directory)] = error
-
-    def set_directory_delay(
-        self,
-        username: str,
-        directory: str,
-        seconds: float,
-    ) -> None:
-        """Sleep `seconds` inside `directory(...)` before returning the registered
-        result. Lets tests exercise fan-out wave deadlines and concurrency caps
-        without mocking time. Default is 0.0 (no sleep)."""
-        self._directory_delays[(username, directory)] = seconds
-
-    def directory(self, username: str, directory: str) -> list[Any]:
-        self.directory_calls.append((username, directory))
-        if self.in_flight_probe is not None:
-            self.in_flight_probe(1)
-        try:
-            delay = self._directory_delays.get((username, directory), 0.0)
-            if delay > 0:
-                time.sleep(delay)
-            if self.directory_error is not None:
-                raise self.directory_error
-            directory_error = self._directory_errors.get((username, directory))
-            if directory_error is not None:
-                raise directory_error
-            return copy.deepcopy(self._directories.get((username, directory), []))
-        finally:
-            if self.in_flight_probe is not None:
-                self.in_flight_probe(-1)
-
-    def set_status(self, username: str, presence: str) -> None:
-        """Configure the presence returned by ``status(username)``.
-
-        ``presence`` is the slskd-api ``UserPresence`` value -- one of
-        ``"Online"``, ``"Away"``, ``"Offline"``.
-        """
-        self._statuses[username] = presence
-
-    def set_status_error(self, username: str, error: Exception) -> None:
-        """Configure ``status(username)`` to raise ``error``."""
-        self._status_errors[username] = error
-
-    def status(self, username: str) -> dict[str, Any]:
-        """Mirror slskd-api ``UsersApi.status`` shape: ``{"presence": ...,
-        "isPrivileged": False}``. Default presence is ``"Online"`` so legacy
-        tests that don't configure status stay green."""
-        self.status_calls.append(username)
-        configured_error = self._status_errors.get(username)
-        if configured_error is not None:
-            raise configured_error
-        presence = self._statuses.get(username, "Online")
-        return {"presence": presence, "isPrivileged": False}
-
-
-@dataclass
-class SearchTextCall:
-    """One slskd ``searches.search_text`` call captured by FakeSlskdSearches."""
-    search_text: str
-    kwargs: dict[str, Any]
-
-
-class FakeSlskdSearches:
-    """Stateful fake for the slskd searches API.
-
-    Drives orchestration tests over `search_for_album` / `_submit_plan_search`:
-    pre-seed canned ``state``, ``responses`` for known search ids; record
-    every ``search_text`` kwargs (especially ``responseLimit``) for later
-    assertion.
-
-    Usage:
-        searches = FakeSlskdSearches()
-        searches.add_search(search_id=1, state="Completed", responses=[...])
-        searches.search_text_id_sequence = [1]   # next call returns id=1
-        # ... drive code under test ...
-        assert searches.search_text_calls[0].kwargs["responseLimit"] == 1000
-    """
-
-    def __init__(self) -> None:
-        self.search_text_calls: list[SearchTextCall] = []
-        self.state_calls: list[tuple[Any, bool]] = []
-        self.responses_calls: list[Any] = []
-        self.delete_calls: list[Any] = []
-        self.stop_calls: list[Any] = []
-        self.search_text_error: Exception | None = None
-        # Per-search override: id -> Exception. Raised from stop() / state()
-        # for that search id. Used to drive the "stop() raises" / "state()
-        # raises" branches without poisoning every search.
-        self._stop_errors: dict[Any, Exception] = {}
-        self._stop_returns: dict[Any, bool] = {}
-        # Each call returns the next id from this list; falls back to a
-        # monotonically incrementing counter once the list is exhausted.
-        self.search_text_id_sequence: list[Any] = []
-        self._next_auto_id = 1
-        # search_id -> {
-        #   "state": str,
-        #   "responses": list[dict],
-        #   "response_count": int,
-        #   "post_stop_state": str | None,
-        #   "post_stop_responses": list[dict] | None,
-        # }
-        self._searches: dict[Any, dict[str, Any]] = {}
-
-    def add_search(
-        self,
-        *,
-        search_id: Any,
-        state: str = "Completed",
-        responses: list[dict[str, Any]] | None = None,
-        response_count: int | None = None,
-        post_stop_state: str | None = None,
-        post_stop_responses: list[dict[str, Any]] | None = None,
-    ) -> None:
-        """Pre-register a canned response set for a search id.
-
-        The watchdog reads ``state_resp["responseCount"]`` to track
-        no-progress. When ``response_count`` is omitted it defaults to
-        ``len(responses)``. Tests that need to simulate a stuck search with
-        non-zero starting responses but no further progress can pass
-        ``response_count`` explicitly.
-
-        ``post_stop_state`` / ``post_stop_responses`` simulate slskd's async
-        cleanup after ``stop()``: the next ``state()`` call after ``stop()``
-        flips the state and ``search_responses()`` flips the responses.
-        Leave both unset to model "slskd hung at cleanup" — state stays
-        InProgress until the watchdog gives up.
-        """
-        self._searches[search_id] = {
-            "state": state,
-            "responses": copy.deepcopy(responses or []),
-            "response_count": response_count if response_count is not None
-                              else len(responses or []),
-            "post_stop_state": post_stop_state,
-            "post_stop_responses": (
-                copy.deepcopy(post_stop_responses)
-                if post_stop_responses is not None else None
-            ),
-            "_stopped": False,
-        }
-
-    def set_response_count(self, search_id: Any, count: int) -> None:
-        """Mutate the responseCount of an already-seeded search.
-
-        Tests model "responses arrive over time" by stepping the count
-        between calls to advance the watchdog clock.
-        """
-        if search_id in self._searches:
-            self._searches[search_id]["response_count"] = count
-
-    def set_state(self, search_id: Any, state: str) -> None:
-        """Mutate the state of an already-seeded search."""
-        if search_id in self._searches:
-            self._searches[search_id]["state"] = state
-
-    def set_stop_error(self, search_id: Any, err: Exception) -> None:
-        self._stop_errors[search_id] = err
-
-    def set_stop_return(self, search_id: Any, value: bool) -> None:
-        self._stop_returns[search_id] = value
-
-    def search_text(self, **kwargs: Any) -> dict[str, Any]:
-        text = kwargs.pop("searchText", "")
-        self.search_text_calls.append(
-            SearchTextCall(search_text=text, kwargs=copy.deepcopy(kwargs)))
-        if self.search_text_error is not None:
-            raise self.search_text_error
-        if self.search_text_id_sequence:
-            search_id = self.search_text_id_sequence.pop(0)
-        else:
-            search_id = self._next_auto_id
-            self._next_auto_id += 1
-        # Default state for an unconfigured id keeps the fake usable
-        # without explicit ``add_search`` for one-shot tests.
-        self._searches.setdefault(search_id, {
-            "state": "Completed",
-            "responses": [],
-        })
-        return {"id": search_id}
-
-    def _maybe_apply_post_stop_flip(self, cfg: dict[str, Any]) -> None:
-        """Apply the post-stop state/response flip on the first observation
-        after ``stop()``. Called from both ``state()`` and
-        ``search_responses()`` because slskd's real async cleanup commits
-        the response store regardless of which endpoint we poll. Pre-#242
-        the flip was state()-gated; that no longer matches the production
-        helper which polls responses directly.
-        """
-        if cfg.get("_stopped") and cfg.get("post_stop_state") is not None:
-            cfg["state"] = cfg["post_stop_state"]
-            cfg["post_stop_state"] = None  # only flip once
-            if cfg.get("post_stop_responses") is not None:
-                cfg["responses"] = cfg["post_stop_responses"]
-                cfg["response_count"] = len(cfg["responses"])
-                cfg["post_stop_responses"] = None
-
-    def state(self, search_id: Any, _include_responses: bool = False) -> dict[str, Any]:
-        self.state_calls.append((search_id, _include_responses))
-        cfg = self._searches.get(search_id)
-        if cfg is None:
-            return {
-                "id": search_id,
-                "state": "Completed",
-                "isComplete": True,
-                "responseCount": 0,
-            }
-        self._maybe_apply_post_stop_flip(cfg)
-        return {
-            "id": search_id,
-            "state": cfg["state"],
-            "isComplete": True,
-            "responseCount": cfg.get("response_count", 0),
-        }
-
-    def search_responses(self, search_id: Any) -> list[dict[str, Any]]:
-        self.responses_calls.append(search_id)
-        cfg = self._searches.get(search_id)
-        if cfg is None:
-            return []
-        self._maybe_apply_post_stop_flip(cfg)
-        return copy.deepcopy(cfg["responses"])
-
-    def stop(self, search_id: Any) -> bool:
-        self.stop_calls.append(search_id)
-        # Mark _stopped FIRST. This models slskd's real behaviour: even
-        # when the wrapper call raises (e.g., transport error mid-PUT),
-        # the server may have already received the cancel and started its
-        # async cleanup. Tests that simulate stop() errors still need the
-        # post_stop_state flip to fire so the post-cancel wait can exit
-        # without spinning the deadline budget down to zero.
-        cfg = self._searches.get(search_id)
-        if cfg is not None:
-            cfg["_stopped"] = True
-        if search_id in self._stop_errors:
-            raise self._stop_errors[search_id]
-        return self._stop_returns.get(search_id, True)
-
-    def delete(self, search_id: Any) -> None:
-        self.delete_calls.append(search_id)
-
-
-class FakeSlskdAPI:
-    """In-memory fake for slskd API clients used by download tests."""
-
-    def __init__(
-        self,
-        *,
-        downloads: list[dict[str, Any]] | None = None,
-        download_snapshots: list[list[dict[str, Any]]] | None = None,
-    ) -> None:
-        self.transfers = FakeSlskdTransfers(self)
-        self.users = FakeSlskdUsers()
-        self.searches = FakeSlskdSearches()
-        self._downloads = copy.deepcopy(downloads or [])
-        self._download_snapshots = [
-            copy.deepcopy(snapshot) for snapshot in (download_snapshots or [])
-        ]
-
-    def set_downloads(self, downloads: list[dict[str, Any]]) -> None:
-        self._downloads = copy.deepcopy(downloads)
-        self._download_snapshots = []
-
-    def queue_download_snapshots(self, *snapshots: list[dict[str, Any]]) -> None:
-        self._download_snapshots.extend(copy.deepcopy(list(snapshots)))
-
-    def add_transfer(
-        self,
-        *,
-        username: str,
-        directory: str,
-        filename: str,
-        id: str,
-        state: str | None = None,
-        size: int | None = None,
-        bytesTransferred: int | None = None,
-        **extra: Any,
-    ) -> None:
-        group = self._find_or_create_group(username)
-        directory_row = self._find_or_create_directory(group, directory)
-        transfer: dict[str, Any] = {"filename": filename, "id": id}
-        if state is not None:
-            transfer["state"] = state
-        if size is not None:
-            transfer["size"] = size
-        if bytesTransferred is not None:
-            transfer["bytesTransferred"] = bytesTransferred
-        transfer.update(extra)
-        directory_row.setdefault("files", []).append(transfer)
-
-    def _next_download_snapshot(self) -> list[dict[str, Any]]:
-        if self._download_snapshots:
-            self._downloads = self._download_snapshots.pop(0)
-        return copy.deepcopy(self._downloads)
-
-    def _find_transfer(self, username: str, transfer_id: str) -> dict[str, Any] | None:
-        for group in self._downloads:
-            if group.get("username") not in (None, "", username):
-                continue
-            for directory in group.get("directories", []):
-                for transfer in directory.get("files", []):
-                    if transfer.get("id") == transfer_id:
-                        return copy.deepcopy(transfer)
-        return None
-
-    def remove_transfer(self, *, username: str, id: str) -> None:
-        for group in self._downloads:
-            if group.get("username") not in (None, "", username):
-                continue
-            for directory in group.get("directories", []):
-                directory["files"] = [
-                    transfer for transfer in directory.get("files", [])
-                    if transfer.get("id") != id
-                ]
-
-    def _find_or_create_group(self, username: str) -> dict[str, Any]:
-        for group in self._downloads:
-            if group.get("username") == username:
-                return group
-        group = {"username": username, "directories": []}
-        self._downloads.append(group)
-        return group
-
-    @staticmethod
-    def _find_or_create_directory(
-        group: dict[str, Any],
-        directory: str,
-    ) -> dict[str, Any]:
-        for row in group.setdefault("directories", []):
-            if row.get("directory") == directory:
-                return row
-        row = {"directory": directory, "files": []}
-        group["directories"].append(row)
-        return row
-
-
-class FakeYTMusic:
-    """In-memory fake for the slice of ``ytmusicapi.YTMusic`` the YouTube
-    album resolver uses (``search`` + ``get_album``).
-
-    Mirrors ``FakeSlskdAPI`` in shape: pre-seed canned per-query results,
-    queue one-shot exceptions to simulate upstream failures, and record
-    every call so service tests can assert N+1 fan-out shape.
-
-    Exceptions to inject mirror the real ``ytmusicapi`` boundary:
-    ``YTMusicServerError`` (4xx/5xx upstream), ``YTMusicUserError``
-    (malformed query), ``requests.Timeout`` / ``requests.ConnectionError``
-    (transport layer), and ``KeyError`` (parser drift / library version
-    skew).
-    """
-
-    def __init__(self) -> None:
-        self._search_results: dict[str, list[dict[str, Any]]] = {}
-        self._album_results: dict[str, dict[str, Any]] = {}
-        self._search_errors: dict[str, Exception] = {}
-        self._album_errors: dict[str, Exception] = {}
-        self.search_calls: list[dict[str, Any]] = []
-        self.get_album_calls: list[dict[str, Any]] = []
-
-    def set_search(
-        self,
-        query: str,
-        results: list[dict[str, Any]],
-    ) -> None:
-        """Configure ``search(query, ...)`` to return ``results``."""
-        self._search_results[query] = copy.deepcopy(results)
-
-    def set_album(
-        self,
-        browseId: str,
-        response: dict[str, Any],
-    ) -> None:
-        """Configure ``get_album(browseId)`` to return ``response``."""
-        self._album_results[browseId] = copy.deepcopy(response)
-
-    def set_search_error(
-        self,
-        query: str,
-        error: Exception,
-    ) -> None:
-        """Queue a single exception to raise on the next ``search(query, ...)``
-        call. One-shot: subsequent matching calls fall through to the canned
-        result (or the empty default)."""
-        self._search_errors[query] = error
-
-    def set_album_error(
-        self,
-        browseId: str,
-        error: Exception,
-    ) -> None:
-        """Queue a single exception to raise on the next
-        ``get_album(browseId)``. One-shot: subsequent matching calls fall
-        through to the canned result (or the default unconfigured-raise)."""
-        self._album_errors[browseId] = error
-
-    def search(
-        self,
-        query: str,
-        filter: str | None = None,
-        scope: str | None = None,
-        limit: int = 20,
-        ignore_spelling: bool = False,
-    ) -> list[dict[str, Any]]:
-        """Mirror ``ytmusicapi.YTMusic.search``. Returns the canned list for
-        ``query`` if configured, else an empty list (the real library returns
-        ``[]`` for "no hits" — not an error)."""
-        self.search_calls.append({
-            "query": query,
-            "filter": filter,
-            "scope": scope,
-            "limit": limit,
-            "ignore_spelling": ignore_spelling,
-        })
-        queued_error = self._search_errors.pop(query, None)
-        if queued_error is not None:
-            raise queued_error
-        return copy.deepcopy(self._search_results.get(query, []))
-
-    def get_album(self, browseId: str) -> dict[str, Any]:
-        """Mirror ``ytmusicapi.YTMusic.get_album``. Returns the canned dict
-        for ``browseId`` if configured. Unconfigured browseIds raise
-        ``YTMusicServerError`` — matching the real library's behavior for
-        a non-existent album."""
-        self.get_album_calls.append({"browseId": browseId})
-        queued_error = self._album_errors.pop(browseId, None)
-        if queued_error is not None:
-            raise queued_error
-        if browseId not in self._album_results:
-            # Lazy import — ytmusicapi is a heavy import; only pull it in
-            # when the fake actually needs to raise the real exception type
-            # so tests calling other methods don't pay for it.
-            from ytmusicapi.exceptions import YTMusicServerError
-            raise YTMusicServerError(
-                f"Server returned HTTP 400: no album for browseId={browseId!r}"
-            )
-        return copy.deepcopy(self._album_results[browseId])
-
-    @staticmethod
-    def make_album_fixture(
-        audio_playlist_id: str | None,
-        title: str,
-        artists: list[dict[str, Any]],
-        year: str | None,
-        tracks: list[dict[str, Any]],
-        other_versions: list[dict[str, Any]] | None = None,
-    ) -> dict[str, Any]:
-        """Synthesize a ``get_album()``-shaped dict for use in tests.
-
-        Mirrors the canonical ``ytmusicapi.YTMusic.get_album`` return shape
-        per the library's documented schema. The real response does not
-        carry the album's own ``browseId`` — callers already know it
-        because they passed it to ``get_album()`` — so the fixture omits
-        it too. Pair this with ``set_album(browse_id, fixture)`` so the
-        fake knows which browseId to return it for. Tests that only care
-        about a subset of fields can still call ``set_album`` with
-        whatever subset they need — this helper is for the happy-path
-        / integration-slice cases where the service's parsing code reads
-        many of these fields.
-
-        Track-level keys: ``videoId``, ``title``, ``artists``, ``album``,
-        ``duration``, ``duration_seconds``, ``trackNumber``, ``isAvailable``,
-        ``isExplicit``, ``likeStatus``, ``thumbnails``, ``feedbackTokens``,
-        ``creditsBrowseId``.
-
-        Other-version keys: ``browseId``, ``title``, ``artists``, ``year``,
-        ``thumbnails``, ``isExplicit``.
-        """
-        total_seconds = sum(
-            int(t.get("duration_seconds", 0) or 0) for t in tracks
-        )
-        # Format total duration as M:SS or H:MM:SS (best-effort, matches
-        # YTMusic's typical "30:14" or "1:02:14" formats).
-        if total_seconds >= 3600:
-            hours = total_seconds // 3600
-            minutes = (total_seconds % 3600) // 60
-            seconds = total_seconds % 60
-            duration_str = f"{hours}:{minutes:02d}:{seconds:02d}"
-        else:
-            minutes = total_seconds // 60
-            seconds = total_seconds % 60
-            duration_str = f"{minutes}:{seconds:02d}"
-        return {
-            "title": title,
-            "type": "Album",
-            "thumbnails": [],
-            "description": None,
-            "artists": copy.deepcopy(artists),
-            "year": year,
-            "trackCount": len(tracks),
-            "duration": duration_str,
-            "duration_seconds": total_seconds,
-            "audioPlaylistId": audio_playlist_id,
-            "tracks": copy.deepcopy(tracks),
-            "other_versions": copy.deepcopy(other_versions or []),
-        }
-
-
-# ---------------------------------------------------------------------------
-# Mirror-adapter lookup fakes (test-fidelity Rule B).
-#
-# ``web/mb.py::get_release`` and ``web/discogs.py::get_release`` RAISE
-# ``urllib.error.HTTPError(404)`` when the id is absent — they NEVER return
-# ``None``. A test that fakes a miss with ``lambda mbid: None`` simulates a
-# code path production never produces. This was the round-1 P0 of the
-# YT-resolver PR: ``_resolve_mb_group`` expected ``None`` on 404, the real
-# adapter raised, and every test used ``lambda: None`` so the production
-# crash never surfaced. These fakes mirror the real exception contract:
-# seed known releases; any un-seeded id raises ``HTTPError(404)`` by
-# default. See ``.claude/rules/test-fidelity.md`` § "Rule B".
-# ---------------------------------------------------------------------------
-
-
-def http_error(
-    code: int,
-    *,
-    url: str = "http://mirror.test/x",
-    msg: str | None = None,
-) -> urllib.error.HTTPError:
-    """Construct a real ``urllib.error.HTTPError`` for ``code``.
-
-    The canonical way to make a fake adapter raise exactly what the live
-    MB / Discogs mirror raises on a 4xx/5xx. ``HTTPError`` is a subclass
-    of ``urllib.error.URLError``, so callers that catch either type see it.
-    """
-    return urllib.error.HTTPError(
-        url, code, msg or f"HTTP {code}", email.message.Message(), None,
-    )
-
-
-class FakeMBLookup:
-    """Callable stand-in for ``web.mb.get_release`` with the REAL exception
-    contract.
-
-    Production ``web.mb.get_release(mbid)`` raises
-    ``urllib.error.HTTPError(code=404)`` for an absent MBID and returns a
-    slim release dict for a present one — it NEVER returns ``None``. Faking
-    a miss with ``lambda mbid: None`` is the test-fidelity Rule B
-    anti-pattern: it exercises a ``None`` branch production cannot reach
-    while hiding the exception branch production actually takes.
-
-    Usage::
-
-        mb = FakeMBLookup()
-        mb.set_release("rel-1", {"id": "rel-1", "title": "X", ...})
-        resolve(..., mb_get_release=mb)   # hit -> dict
-        # any other mbid -> raises HTTPError(404), exactly like production
-
-    Pass ``raises_on_404=False`` only to model an adapter variant that
-    genuinely returns ``None`` on miss (rare — document why at the call
-    site). Accepts the ``fresh=`` kwarg the production callable takes.
-    """
-
-    def __init__(
-        self,
-        releases: dict[str, dict[str, Any]] | None = None,
-        *,
-        raises_on_404: bool = True,
-    ) -> None:
-        self._releases: dict[str, dict[str, Any]] = dict(releases or {})
-        self._raises_on_404 = raises_on_404
-        self.calls: list[str] = []
-
-    def set_release(
-        self, identifier: str, payload: dict[str, Any],
-    ) -> "FakeMBLookup":
-        """Seed a hit for ``identifier``. Returns self for chaining."""
-        self._releases[identifier] = copy.deepcopy(payload)
-        return self
-
-    def __call__(
-        self, identifier: str, *, fresh: bool = False,
-    ) -> Optional[dict[str, Any]]:
-        self.calls.append(identifier)
-        if identifier in self._releases:
-            return copy.deepcopy(self._releases[identifier])
-        if self._raises_on_404:
-            raise http_error(404, url=f"http://mb.test/release/{identifier}")
-        return None
-
-
-class FakeDiscogsLookup:
-    """Callable stand-in for ``web.discogs.get_release`` with the REAL
-    exception contract — the Discogs analogue of :class:`FakeMBLookup`.
-
-    Production ``web.discogs.get_release(release_id)`` raises
-    ``urllib.error.HTTPError(404)`` for an absent id (the mirror's ``_get``
-    propagates ``urlopen``'s ``HTTPError`` directly — no retry) and never
-    returns ``None``.
-    """
-
-    def __init__(
-        self,
-        releases: dict[str, dict[str, Any]] | None = None,
-        *,
-        raises_on_404: bool = True,
-    ) -> None:
-        self._releases: dict[str, dict[str, Any]] = dict(releases or {})
-        self._raises_on_404 = raises_on_404
-        self.calls: list[str] = []
-
-    def set_release(
-        self, identifier: str, payload: dict[str, Any],
-    ) -> "FakeDiscogsLookup":
-        """Seed a hit for ``identifier``. Returns self for chaining."""
-        self._releases[identifier] = copy.deepcopy(payload)
-        return self
-
-    def __call__(
-        self, identifier: str, *, fresh: bool = False,
-    ) -> Optional[dict[str, Any]]:
-        self.calls.append(identifier)
-        if identifier in self._releases:
-            return copy.deepcopy(self._releases[identifier])
-        if self._raises_on_404:
-            raise http_error(
-                404, url=f"http://discogs.test/releases/{identifier}")
-        return None
 
 
 class FakePipelineDB:
@@ -5708,389 +4850,3 @@ class FakePipelineDBSource:
     def close(self) -> None:
         self.close_calls += 1
 
-
-class FakeCursor:
-    """Minimal DB-API cursor stand-in for raw-SQL seams.
-
-    Pair with :meth:`FakePipelineDB.queue_execute_results` when the
-    code under test goes through ``PipelineDB._execute`` directly
-    (e.g. ``web.overlay.check_pipeline``) — the fake cannot interpret
-    SQL, so the test supplies the rows the query would return.
-
-    Mirrors real psycopg2 consumption semantics (test-fidelity Rule B):
-    ``fetchone`` advances and returns ``None`` when exhausted;
-    ``fetchall`` drains and returns only the remainder. A
-    ``while cur.fetchone():`` consumer must terminate.
-    """
-
-    def __init__(self, rows: list[dict[str, Any]] | None = None) -> None:
-        self._rows = list(rows) if rows else []
-
-    def fetchall(self) -> list[dict[str, Any]]:
-        rows, self._rows = self._rows, []
-        return rows
-
-    def fetchone(self) -> dict[str, Any] | None:
-        return self._rows.pop(0) if self._rows else None
-
-
-class FakeBeetsDB:
-    """In-memory fake for ``lib.beets_db.BeetsDB`` — minimal surface.
-
-    Records state for assertions and lets tests register return values
-    for the query methods used by ``harness/import_one.py`` and a few
-    web routes. Intentionally narrow — extend only when a test surface
-    actually exercises a new method. The real ``BeetsDB`` is the
-    contract; the audit's ``test_fake_only_methods_stay_on_the_allowlist``
-    catches drift.
-
-    Usage:
-        beets = FakeBeetsDB()
-        beets.set_album_exists("mbid-123", False)
-        beets.set_album_ids_for_release("mbid-123", [77])
-        beets.set_album_info(77, AlbumInfo(...))
-        beets.set_item_paths("mbid-123", [(11, "/path/01.flac")])
-        # ... drive code that calls beets.album_exists("mbid-123") ...
-        assert beets.close_calls == 1
-    """
-
-    def __init__(self) -> None:
-        self._album_exists: dict[str, bool] = {}
-        self._album_ids_for_release: dict[str, list[int]] = {}
-        self._album_info: dict[str, Any] = {}
-        self._item_paths: dict[str, list[tuple[int, str]]] = {}
-        self._album_path_by_id: dict[int, str | None] = {}
-        self._release_identities: list[dict[str, Any]] = []
-        # Default return values for unseeded keys — match the real
-        # BeetsDB's "no row" shapes so tests don't crash on missing
-        # explicit seeds.
-        self._album_exists_default = False
-        self._album_ids_default: list[int] = []
-        self._album_info_default: Any = None
-        self._album_path_by_id_default: str | None = None
-        self._item_paths_default: list[tuple[int, str]] = []
-        self.close_calls: int = 0
-        self.album_exists_calls: list[str] = []
-        self.get_album_info_calls: list[str] = []
-        self.get_all_album_ids_for_release_calls: list[str] = []
-        self.get_album_path_by_id_calls: list[int] = []
-        self.get_item_paths_calls: list[str] = []
-        self.check_mbids_calls: list[list[str]] = []
-        self.check_mbids_detail_calls: list[list[str]] = []
-        self.get_album_ids_by_mbids_calls: list[list[str]] = []
-        self.get_tracks_by_mb_release_id_calls: list[str] = []
-        self._tracks_by_release: dict[str, list[dict[str, Any]]] = {}
-        self.get_albums_by_artist_calls: list[tuple[str, str]] = []
-        self._mbid_detail: dict[str, dict[str, Any]] = {}
-        self._albums_by_artist: dict[str, list[dict[str, Any]]] = {}
-        self.list_release_identities_calls: int = 0
-        self._library_albums: list[dict[str, Any]] = []
-        self._album_detail: dict[int, dict[str, Any]] = {}
-        self.search_albums_calls: list[tuple[str, int]] = []
-        self.get_recent_calls: list[int] = []
-        self.get_album_detail_calls: list[int] = []
-        self._locate_queue: list[ReleaseLocation] = []
-        self.locate_calls: list[str] = []
-        self._min_bitrate: dict[str, int | None] = {}
-        self._min_bitrate_default: int | None = None
-        self.get_min_bitrate_calls: list[str] = []
-
-    # --- Seeding helpers ---
-
-    def set_mbid_detail(self, mbid: str, detail: dict[str, Any]) -> None:
-        self._mbid_detail[mbid] = detail
-
-    def set_albums_by_artist(
-        self, name: str, albums: list[dict[str, Any]],
-    ) -> None:
-        self._albums_by_artist[name] = albums
-
-    def set_album_exists(self, release_id: str, value: bool) -> None:
-        self._album_exists[release_id] = value
-
-    def set_album_ids_for_release(
-        self, release_id: str, ids: list[int],
-    ) -> None:
-        self._album_ids_for_release[release_id] = list(ids)
-
-    def set_album_info(self, mb_release_id: str, info: Any) -> None:
-        self._album_info[mb_release_id] = info
-
-    def set_item_paths(
-        self, release_id: str, paths: list[tuple[int, str]],
-    ) -> None:
-        self._item_paths[release_id] = list(paths)
-
-    def set_album_path_by_id(self, album_id: int, path: str | None) -> None:
-        self._album_path_by_id[album_id] = path
-
-    def set_release_identities(self, rows: list[dict[str, Any]]) -> None:
-        self._release_identities = [copy.deepcopy(r) for r in rows]
-
-    def set_tracks_for_release(
-        self, release_id: str, tracks: list[dict[str, Any]],
-    ) -> None:
-        self._tracks_by_release[release_id] = [
-            copy.deepcopy(t) for t in tracks]
-
-    def set_library_albums(self, rows: list[dict[str, Any]]) -> None:
-        """Seed the album table backing ``search_albums``/``get_recent``."""
-        self._library_albums = [copy.deepcopy(r) for r in rows]
-
-    def set_album_detail(self, album_id: int, detail: dict[str, Any]) -> None:
-        self._album_detail[album_id] = copy.deepcopy(detail)
-
-    def set_min_bitrate(self, release_id: str, kbps: int | None) -> None:
-        """Seed a per-release min bitrate. Keys are normalized like
-        production's locate-backed lookup. The release must also be
-        present (album ids seeded, explicit album_exists, or the
-        default) — ``get_min_bitrate`` gates on presence first."""
-        self._min_bitrate[normalize_release_id(release_id)] = kbps
-
-    def queue_locate_results(self, entries: list[ReleaseLocation]) -> None:
-        """Queue ``locate()`` outcomes for before/after-mutation flows.
-
-        Ban-source observes presence around a ``beet remove`` that runs
-        behind a patched subprocess, so the fake's seed stores can't
-        change mid-flow — the queue models the external mutation.
-        Entries are consumed in order; the last entry repeats. An
-        ``exact`` entry with empty selectors gets them auto-filled from
-        the queried id's shape at call time (the locate contract derives
-        selectors from the ID). With an empty queue, ``locate`` is
-        state-derived from the ``set_album_ids_for_release`` store.
-
-        Production-impossible locations are rejected up front: an
-        ``exact`` hit always carries a real int album_id, and ``absent``
-        is always ``(None, ())`` (lib/beets_db.py::locate).
-        """
-        for entry in entries:
-            if entry.kind == "exact":
-                assert (isinstance(entry.album_id, int)
-                        and not isinstance(entry.album_id, bool)), (
-                    f"exact ReleaseLocation needs an int album_id, got "
-                    f"{entry.album_id!r} — production cannot construct this")
-            else:
-                assert entry.album_id is None and entry.selectors == (), (
-                    f"absent ReleaseLocation is always (None, ()), got "
-                    f"({entry.album_id!r}, {entry.selectors!r})")
-        self._locate_queue = list(entries)
-
-    # --- Real-method surface ---
-
-    def _album_ids_lookup(self, release_id: str) -> list[int] | None:
-        """Normalized view of the ``set_album_ids_for_release`` store.
-
-        Production's ``_batch_lookup_album_ids`` runs every input
-        through ``normalize_release_id`` and matches against the stored
-        column value — '0012856590' finds the row stored '12856590'.
-        Returns ``None`` for "release never seeded" vs ``[]`` for
-        "seeded with zero albums".
-        """
-        key = normalize_release_id(release_id)
-        if not key:
-            return None
-        for seeded, ids in self._album_ids_for_release.items():
-            if normalize_release_id(seeded) == key:
-                return ids
-        return None
-
-    def _presence(self, release_id: str) -> bool:
-        """Current presence per the locate seam (issue #121).
-
-        Production answers ``album_exists`` / ``get_min_bitrate`` via
-        ``locate``; the fake's equivalent precedence is: queued locate
-        head (the most explicit model of "current state around an
-        external mutation") → explicit ``set_album_exists`` seed →
-        seeded album ids → the ``_album_exists_default``.
-        """
-        if self._locate_queue:
-            return self._locate_queue[0].kind == "exact"
-        if release_id in self._album_exists:
-            return self._album_exists[release_id]
-        ids = self._album_ids_lookup(release_id)
-        if ids:
-            return True
-        return self._album_exists_default
-
-    def album_exists(self, release_id: str) -> bool:
-        self.album_exists_calls.append(release_id)
-        return self._presence(release_id)
-
-    def check_mbids(self, mbids: list[str]) -> set[str]:
-        self.check_mbids_calls.append(list(mbids))
-        return {mbid for mbid in mbids if self.album_exists(mbid)}
-
-    def check_mbids_detail(
-        self, mbids: list[str],
-    ) -> dict[str, dict[str, Any]]:
-        self.check_mbids_detail_calls.append(list(mbids))
-        return {
-            mbid: copy.deepcopy(self._mbid_detail[mbid])
-            for mbid in mbids if mbid in self._mbid_detail
-        }
-
-    def get_albums_by_artist(
-        self, name: str, mbid: str = "",
-    ) -> list[dict[str, Any]]:
-        self.get_albums_by_artist_calls.append((name, mbid))
-        return copy.deepcopy(self._albums_by_artist.get(name, []))
-
-    def list_release_identities(self) -> list[dict[str, Any]]:
-        self.list_release_identities_calls += 1
-        return [copy.deepcopy(r) for r in self._release_identities]
-
-    def get_all_album_ids_for_release(self, release_id: str) -> list[int]:
-        self.get_all_album_ids_for_release_calls.append(release_id)
-        return self._album_ids_for_release.get(
-            release_id, list(self._album_ids_default))
-
-    def get_album_ids_by_mbids(self, mbids: list[str]) -> dict[str, int]:
-        """Mirror of ``BeetsDB.get_album_ids_by_mbids`` — exact hits only.
-
-        Derives from the ``set_album_ids_for_release`` seed store —
-        shared with ``get_all_album_ids_for_release``, and seeded album
-        ids imply ``album_exists`` — mirroring production's single
-        ``_batch_lookup_album_ids`` seam (issue #121). Inputs and result
-        keys are normalized like production ('0012856590' hits the row
-        seeded '12856590' and the result is keyed '12856590'). First
-        seeded album id wins, matching the exact-hit ``locate`` contract.
-        """
-        self.get_album_ids_by_mbids_calls.append(list(mbids))
-        result: dict[str, int] = {}
-        for mbid in mbids:
-            key = normalize_release_id(mbid)
-            if not key:
-                continue
-            ids = self._album_ids_lookup(key)
-            if ids is None:
-                ids = list(self._album_ids_default)
-            if ids:
-                result[key] = ids[0]
-        return result
-
-    def get_tracks_by_mb_release_id(
-        self, mbid: str,
-    ) -> list[dict[str, Any]] | None:
-        """Mirror of ``BeetsDB.get_tracks_by_mb_release_id``.
-
-        ``None`` only when the release has no exact beets hit — the
-        browse route branches on that. A release seeded with album ids
-        but no tracks returns ``[]`` (production: exact hit always
-        yields a list), so 'album present but tracks None' is not an
-        expressible state.
-        """
-        self.get_tracks_by_mb_release_id_calls.append(mbid)
-        key = normalize_release_id(mbid)
-        for seeded, tracks in self._tracks_by_release.items():
-            if normalize_release_id(seeded) == key:
-                return [copy.deepcopy(t) for t in tracks]
-        if self._album_ids_lookup(key):
-            return []
-        return None
-
-    def search_albums(
-        self, query: str, limit: int = 100,
-    ) -> list[dict[str, Any]]:
-        """Mirror of ``BeetsDB.search_albums`` over the seeded album
-        table: case-insensitive substring on artist/album, ordered by
-        (artist, year, album), LIMIT applied after ordering."""
-        self.search_albums_calls.append((query, limit))
-        q = query.lower()
-        hits = [
-            a for a in self._library_albums
-            if q in str(a.get("artist", "")).lower()
-            or q in str(a.get("album", "")).lower()
-        ]
-        hits.sort(key=lambda a: (
-            str(a.get("artist", "")), a.get("year") or 0,
-            str(a.get("album", ""))))
-        return [copy.deepcopy(a) for a in hits[:limit]]
-
-    def get_recent(self, limit: int = 50) -> list[dict[str, Any]]:
-        """Mirror of ``BeetsDB.get_recent`` — ``added`` DESC, NULLs last
-        (SQLite DESC ordering), LIMIT applied after ordering."""
-        self.get_recent_calls.append(limit)
-        rows = sorted(
-            self._library_albums,
-            key=lambda a: (a.get("added") is not None, a.get("added") or 0),
-            reverse=True)
-        return [copy.deepcopy(a) for a in rows[:limit]]
-
-    def get_album_detail(self, album_id: int) -> dict[str, Any] | None:
-        """Mirror of ``BeetsDB.get_album_detail`` — None when missing."""
-        self.get_album_detail_calls.append(album_id)
-        detail = self._album_detail.get(album_id)
-        return copy.deepcopy(detail) if detail is not None else None
-
-    @staticmethod
-    def _selectors_for(release_id: str) -> tuple[str, ...]:
-        """Selector tuple for an exact hit, derived from the id shape —
-        mirrors the dispatch in ``BeetsDB.locate`` (issue #121)."""
-        key = normalize_release_id(release_id)
-        if detect_release_source(key) == "discogs":
-            return (f"discogs_albumid:{key}", f"mb_albumid:{key}")
-        return (f"mb_albumid:{key}",)
-
-    def locate(self, release_id: str) -> ReleaseLocation:
-        """Mirror of ``BeetsDB.locate`` — the issue-#121 presence seam.
-
-        Queued results (``queue_locate_results``) win; otherwise the
-        answer is state-derived: exact hit iff the release has seeded
-        album ids, with selectors derived from the id shape.
-        """
-        self.locate_calls.append(release_id)
-        if self._locate_queue:
-            entry = (self._locate_queue.pop(0)
-                     if len(self._locate_queue) > 1
-                     else self._locate_queue[0])
-            if entry.kind == "exact" and not entry.selectors:
-                return ReleaseLocation(
-                    kind="exact", album_id=entry.album_id,
-                    selectors=self._selectors_for(release_id))
-            return entry
-        key = normalize_release_id(release_id)
-        ids = self._album_ids_lookup(key)
-        if ids:
-            return ReleaseLocation(
-                kind="exact", album_id=ids[0],
-                selectors=self._selectors_for(key))
-        return ReleaseLocation(kind="absent", album_id=None, selectors=())
-
-    def get_min_bitrate(self, mb_release_id: str) -> int | None:
-        """Mirror of ``BeetsDB.get_min_bitrate`` (kbps; None = no row).
-
-        Production resolves presence through ``locate`` first and
-        returns None for an absent release, so the fake gates on
-        ``_presence`` before consulting the (normalized-key) seed
-        store / ``_min_bitrate_default``.
-        """
-        self.get_min_bitrate_calls.append(mb_release_id)
-        if not self._presence(mb_release_id):
-            return None
-        key = normalize_release_id(mb_release_id)
-        return self._min_bitrate.get(key, self._min_bitrate_default)
-
-    def get_album_info(
-        self, mb_release_id: str, _cfg: Any = None,
-    ) -> Any:
-        self.get_album_info_calls.append(mb_release_id)
-        return self._album_info.get(mb_release_id, self._album_info_default)
-
-    def get_item_paths(self, release_id: str) -> list[tuple[int, str]]:
-        self.get_item_paths_calls.append(release_id)
-        return self._item_paths.get(
-            release_id, list(self._item_paths_default))
-
-    def get_album_path_by_id(self, album_id: int) -> str | None:
-        self.get_album_path_by_id_calls.append(album_id)
-        return self._album_path_by_id.get(
-            album_id, self._album_path_by_id_default)
-
-    def close(self) -> None:
-        self.close_calls += 1
-
-    def __enter__(self) -> "FakeBeetsDB":
-        return self
-
-    def __exit__(self, *_args: object) -> None:
-        self.close()
